requirements.txt.txt
streamlit
pandas
langchain
langchain_groq
python-dotenv
openai
openpyxl